pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "code_analyzer"
version = "0.1.0"
description = "AI-powered code analysis tool"
requires-python = ">=3.8"
dependencies = [
    "loguru>=0.7.0",
    "crewai>=0.1.0",
    "python-dotenv>=1.0.0",
    "openai>=1.3.0",
    "pendulum>=2.0.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0"
]

[project.optional-dependencies]
test = [
    "pytest-cov>=4.1.0",
]
